pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "f5r-tts"
dynamic = ["version"]
description = "F5R-TTS: Improving Flow Matching based Text-to-Speech with Group Relative Policy Optimization"
readme = "README.md"
license = {text = "MIT License"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "bitsandbytes>0.37.0",
    "cached_path",
    "click",
    "datasets",
    "einops",
    "ema_pytorch>=0.5.2",
    "funasr",
    "gradio>=3.45.2",
    "jieba",
    "librosa",
    "matplotlib",
    "numpy<=1.26.4",
    "python-dotenv",
    "pydub",
    "pypinyin",
    "safetensors",
    "soundfile",
    "tomli",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchdiffeq",
    "tqdm>=4.65.0",
    "transformers",
    "transformers_stream_generator",
    "vocos",
    "wandb",
    "wespeaker",
    "x_transformers>=1.31.14",
    "megatron-core>=0.8.0",
    "verl>=0.5.0",
    "huggingface_hub>=0.23.0",
]

[project.optional-dependencies]
eval = [
    "faster_whisper==0.10.1",
    "funasr",
    "jiwer",
    "modelscope",
    "zhconv",
    "zhon",
]

[project.urls]
Homepage = "https://github.com/shaw0fr/F5R-TTS"

